forgot to rename test files [no ci]
